github/gh-auth-types/app-install-flow/requirements.txt
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
idna==3.10
pycparser==2.23
requests==2.32.5
PyJWT==2.10.1
urllib3==2.5.0
